+++ java/src/test/java/io/github/cyfko/dverify/DatabaseTest.java
@@ -13,7 +13,6 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 
 import javax.sql.DataSource;
 import java.io.IOException;
-import java.time.Duration;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -63,9 +62,8 @@ public abstract class DatabaseTest {
     @EnumSource(value = TokenMode.class)
     public void sign_method_with_valid_data_should_returns_token(TokenMode mode) throws JsonEncodingException {
         UserData data = new UserData("john.doe@example.com");
-        Duration duration = Duration.ofHours(2);
 
-        String token = signer.sign(data, duration, mode);
+        String token = signer.sign(data, 60, mode, mode.name().hashCode());
 
         assertNotNull(token);
         assertFalse(token.isEmpty());
@@ -75,27 +73,24 @@ public abstract class DatabaseTest {
     @EnumSource(value = TokenMode.class)
     public void sign_method_with_invalid_data_should_throws_exception(TokenMode mode) {
         Object invalidData = null; // Simulating invalid data
-        Duration duration = Duration.ofHours(2);
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, mode));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, 60, mode, mode.name().hashCode() + 1));
     }
 
     @ParameterizedTest()
     @EnumSource(value = TokenMode.class)
     public void sign_method_with_expired_duration_should_throws_exception(TokenMode mode) {
         UserData data = new UserData("john.doe@example.com");
-        Duration duration = Duration.ofMinutes(-5); // Negative duration
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, duration, mode));
+        Assertions.assertThrows(IllegalArgumentException.class, () -> signer.sign(data, -5, mode, mode.name().hashCode() + 2));
     }
 
     @ParameterizedTest()
     @EnumSource(value = TokenMode.class)
     public void sign_valid_data_should_returns_token(TokenMode mode) throws JsonEncodingException {
         UserData data = new UserData("john.doe@example.com");
-        Duration duration = Duration.ofHours(2);
 
-        String token = signer.sign(data, duration, mode);
+        String token = signer.sign(data, 60, mode, mode.name().hashCode() + 3);
 
         assertNotNull(token, "JWT should not be null");
         assertFalse(token.isEmpty(), "JWT should not be empty");
@@ -105,17 +100,16 @@ public abstract class DatabaseTest {
     @EnumSource(value = TokenMode.class)
     public void sign_invalid_data_should_throws_exception(TokenMode mode) {
         Object invalidData = null; // Simulating invalid data
-        Duration duration = Duration.ofHours(2);
 
-        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, duration, mode));
+        assertThrows(IllegalArgumentException.class, () -> signer.sign(invalidData, 60, mode, mode.name().hashCode() + 4));
     }
 
     @ParameterizedTest()
     @EnumSource(value = TokenMode.class)
-    public void verify_valid_token_should_returns_payload(TokenMode mode) throws InterruptedException {
+    public void verify_valid_token_should_returns_payload(TokenMode mode) {
         UserData data = new UserData("john.doe@example.com");
-        String token = signer.sign(data, Duration.ofHours(2), mode); // Generate a valid token
-        Thread.sleep(5000); // Wait 5 secs to ensure that the keys has been propagated to kafka
+        String token = signer.sign(data, 600, mode, mode.name().hashCode() + 5); // Generate a valid token
+        //Thread.sleep(5000); // Wait 5 secs to ensure that the keys has been propagated to database
 
         UserData verifiedData = verifier.verify(token, UserData.class);
 
@@ -134,8 +128,8 @@ public abstract class DatabaseTest {
     @EnumSource(value = TokenMode.class)
     public void verify_expired_token_should_throws_exception(TokenMode mode) throws InterruptedException {
         UserData data = new UserData("john.doe@example.com");
-        String token = signer.sign(data, Duration.ofMillis(1), mode); // Token with short duration
-        Thread.sleep(10); // Wait for the token to expire
+        String token = signer.sign(data, 1, mode, mode.name().hashCode() + 6); // Token with short duration
+        Thread.sleep(3000); // Wait 3 seconds for the token to expire
 
         assertThrows(DataExtractionException.class, () -> verifier.verify(token, UserData.class));
     }